+++ app/views/profiles/show.html.haml
@@ -17,3 +17,5 @@
     %h4.profile-stat Follow
     %h4.profile-stat Follower
     %h4.profile-stat Posts
+
+= javascript_pack_tag 'profile'
